# Conflicts:
#	FreeRTOS/Demo/RISC-V-Qemu-sifive_e-Eclipse-GCC/freedom-e-sdk/libwrap/sys/printf.c
#	FreeRTOS/Demo/RISC-V-Qemu-sifive_e-Eclipse-GCC/freedom-e-sdk/libwrap/sys/puts.h

+++ FreeRTOS/Demo/RISC-V_Galois_P1/bsp/xilinx/common/xbasic_types.h
@@ -113,7 +113,7 @@ typedef unsigned char	Xuint8;		/**< unsigned 8-bit */
 typedef char		Xint8;		/**< signed 8-bit */
 typedef unsigned short	Xuint16;	/**< unsigned 16-bit */
 typedef short		Xint16;		/**< signed 16-bit */
-typedef unsigned long	Xuint32;	/**< unsigned 32-bit */
+typedef unsigned int	Xuint32;	/**< unsigned 32-bit */
 typedef long		Xint32;		/**< signed 32-bit */
 typedef float		Xfloat32;	/**< 32-bit floating point */
 typedef double		Xfloat64;	/**< 64-bit double precision FP */
